# Use ilocs end-to-end for knowledge graph ranking